TippingTest.jl

A Julia implementation of the tipping classification test proposed by Börner & Dijsktra.

To tip or not to tip

Given a timeseries of a system observable $x(t)$ and corresponding forcing $F(t)$ over a time interval $\tau = [t_0, t_1]$, the test applies the tipping definition by Börner & Dijsktra to determine whether the system tips during $\tau$.

Example

Consider Stommel's ocean box model (Stommel, 1961) of the thermohaline circulation, given in non-dimensional form as a 2-dimensional system of ODEs:

\[\begin{align} \dot{x} &= -\alpha(x-1) - x(1+\mu(x-y)^2) \nonumber \\ \dot{y} &= F_a - y(1+\mu(x-y)^2) \end{align}\]

Here $x$ and $y$ represent the meridional temperature and salinity difference between the southern and northern box, respectively. The freshwater flux $F_a$ is the main forcing parameter; typical values for the model parameters are $\alpha=360$ and $\mu=6.25$.

The relevant observable is the overturning strength $q = 1 + \mu (x-y)^2$. We are interested in the response to the forcing parameter $F_a$.

API

TippingTest.LinearNullModel — Type

LinearNullModel

Abstract supertype for a linear null model defined by a differential equation and a set of parameters.

TippingTest.RelaxNullModel — Type

RelaxNullModel <: LinearNullModel

A simple linear dynamical system with inertia $m$ that responds to the forcing $F$ with sensitivity $lpha$, following the differential equation

\[m \dot y = -(y - y_0) + \alpha F(t) \.\]

Here $y_0$ is the reference equilibrium at $F=0$.

Reference

Börner, R. and Dijkstra, H. A.: To tip or not to tip, EGU General Assembly 2026, Vienna, Austria, 3–8 May 2026, EGU26-18498, https://doi.org/10.5194/egusphere-egu26-18498, 2026

This work is funded by the ClimTip project.

Package maintainer: Reyk Börner (r.borner@uu.nl)