API

TippingTest.LinearNullModel — Type

LinearNullModel

Abstract supertype for a linear null model defined by a differential equation and a set of parameters.

TippingTest.RelaxNullModel — Type

RelaxNullModel <: LinearNullModel

A simple linear dynamical system with inertia $m$ that responds to the forcing $F$ with sensitivity $lpha$, following the differential equation

\[m \dot y = -(y - y_0) + \alpha F(t) \.\]

Here $y_0$ is the reference state at $F=0$.

Fields

y0: Reference state

params: Parameter vector [alpha, m]

TippingTest.TippingTestResult — Type

TippingTestResult

Output of the tipping_test method.

Fields

nullmodel: The fitted linear null model

linear_response: The response time series predicted by the linear null model

residual: The nonlinear residual response, i.e. the component of the observable not explainable by linear response

state_dependence: Time series of the derivative dR/dx, where R is the residual and x the linear response

forcing_dependence: Time series of the derivative dR/dF, where R is the residual and F the forcing

TippingTest.fit_nullmodel — Method

fit_nullmodel(observable, forcing, p_guess; kwargs...)

Fits a LinearNullModel to the observable time series given a forcing time series.

Starting from a parameter guess p_guess, fitting is performed as a weighted least-squares fit using the Optim.jl package. Relative weights can be specified for the time intervals t_A (pre-transition phase), t_tr (transition phase) and t_B (post-transition phase) using the weights keyword argument. The default is weights=[1,0,0], i.e. the fitting is only performed over t_A.

Returns the timeseries of the fitted null model and the optimal null model parameter vector.

Keyword arguments

model = RelaxNullModel: Type of LinearNullModel

time = 0:length(observable)-1: Time range of observable (and forcing)

t_tr = (t_s, t_e): Tuple containing the start (t_s) and end (t_e) time of the transition phase t_tr

y0 = observable[1]: Reference state of null model

weights = [1, 0, 0]: Relative fitting weights within the time intervals [t_A, t_tr, t_B]

TippingTest.tipping_test — Method

tipping_test(observable, forcing, time, t_tr; kwargs...)

Runs the tipping analysis for a given observable timeseries and corresponding forcing timeseries.

The time argument specifies the set of time points at which the observable and forcing data are known (Note: both observable and forcing vectors must have the same length and refer to the same time points).

With the Tuple t_tr = (t_s, t_e), the user specifies the start (t_s) and end (t_e) times of the transition phase t_tr. This phase should contain the majority of the shift from regime A to regime B.

Keyword arguments

model: Instance of a LinearNullModel with initial guess for its parameters p

weights: Vector of fitting weights applied to the time intervals [t_A, t_tr, t_B]